Update package to be grakn.client.GraknClient
